+++ lib/model/actions/entry_actions.dart
@@ -7,10 +7,11 @@ enum EntryAction {
   addShortcut,
   copyToClipboard,
   delete,
-  export,
-  info,
+  convert,
   print,
   rename,
+  copy,
+  move,
   share,
   toggleFavourite,
   // raster
@@ -31,25 +32,32 @@ enum EntryAction {
 }
 
 class EntryActions {
-  static const inApp = [
-    EntryAction.info,
-    EntryAction.toggleFavourite,
+  static const topLevel = [
     EntryAction.share,
-    EntryAction.delete,
+    EntryAction.edit,
     EntryAction.rename,
-    EntryAction.export,
-    EntryAction.addShortcut,
-    EntryAction.copyToClipboard,
-    EntryAction.print,
+    EntryAction.delete,
+    EntryAction.copy,
+    EntryAction.move,
+    EntryAction.toggleFavourite,
     EntryAction.viewSource,
     EntryAction.rotateScreen,
   ];
 
-  static const externalApp = [
-    EntryAction.edit,
+  static const export = [
+    EntryAction.convert,
+    EntryAction.addShortcut,
+    EntryAction.copyToClipboard,
+    EntryAction.print,
     EntryAction.open,
-    EntryAction.setAs,
     EntryAction.openMap,
+    EntryAction.setAs,
+  ];
+
+  static const exportExternal = [
+    EntryAction.open,
+    EntryAction.openMap,
+    EntryAction.setAs,
   ];
 
   static const pageActions = [
@@ -68,14 +76,16 @@ extension ExtraEntryAction on EntryAction {
         return context.l10n.entryActionCopyToClipboard;
       case EntryAction.delete:
         return context.l10n.entryActionDelete;
-      case EntryAction.export:
-        return context.l10n.entryActionExport;
-      case EntryAction.info:
-        return context.l10n.entryActionInfo;
+      case EntryAction.convert:
+        return context.l10n.entryActionConvert;
       case EntryAction.print:
         return context.l10n.entryActionPrint;
       case EntryAction.rename:
         return context.l10n.entryActionRename;
+      case EntryAction.copy:
+        return context.l10n.collectionActionCopy;
+      case EntryAction.move:
+        return context.l10n.collectionActionMove;
       case EntryAction.share:
         return context.l10n.entryActionShare;
       case EntryAction.toggleFavourite:
@@ -133,14 +143,16 @@ extension ExtraEntryAction on EntryAction {
         return AIcons.clipboard;
       case EntryAction.delete:
         return AIcons.delete;
-      case EntryAction.export:
-        return AIcons.saveAs;
-      case EntryAction.info:
-        return AIcons.info;
+      case EntryAction.convert:
+        return AIcons.convert;
       case EntryAction.print:
         return AIcons.print;
       case EntryAction.rename:
         return AIcons.rename;
+      case EntryAction.copy:
+        return AIcons.copy;
+      case EntryAction.move:
+        return AIcons.move;
       case EntryAction.share:
         return AIcons.share;
       case EntryAction.toggleFavourite:
@@ -158,10 +170,13 @@ extension ExtraEntryAction on EntryAction {
         return AIcons.vector;
       // external
       case EntryAction.edit:
+        return AIcons.edit;
       case EntryAction.open:
+        return AIcons.openOutside;
       case EntryAction.openMap:
+        return AIcons.map;
       case EntryAction.setAs:
-        return null;
+        return AIcons.setAs;
       // platform
       case EntryAction.rotateScreen:
         return AIcons.rotateScreen;