+++ lib/widgets/collection/entry_set_action_delegate.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:aves/app_mode.dart';
 import 'package:aves/model/actions/entry_set_actions.dart';
@@ -8,9 +7,7 @@ import 'package:aves/model/device.dart';
 import 'package:aves/model/entry.dart';
 import 'package:aves/model/entry_metadata_edition.dart';
 import 'package:aves/model/favourites.dart';
-import 'package:aves/model/filters/album.dart';
 import 'package:aves/model/filters/filters.dart';
-import 'package:aves/model/highlight.dart';
 import 'package:aves/model/query.dart';
 import 'package:aves/model/selection.dart';
 import 'package:aves/model/source/analysis_controller.dart';
@@ -18,10 +15,8 @@ import 'package:aves/model/source/collection_lens.dart';
 import 'package:aves/model/source/collection_source.dart';
 import 'package:aves/services/common/image_op_events.dart';
 import 'package:aves/services/common/services.dart';
-import 'package:aves/services/media/enums.dart';
 import 'package:aves/theme/durations.dart';
 import 'package:aves/utils/mime_utils.dart';
-import 'package:aves/widgets/collection/collection_page.dart';
 import 'package:aves/widgets/common/action_mixins/entry_editor.dart';
 import 'package:aves/widgets/common/action_mixins/feedback.dart';
 import 'package:aves/widgets/common/action_mixins/permission_aware.dart';
@@ -29,8 +24,6 @@ import 'package:aves/widgets/common/action_mixins/size_aware.dart';
 import 'package:aves/widgets/common/extensions/build_context.dart';
 import 'package:aves/widgets/dialogs/add_shortcut_dialog.dart';
 import 'package:aves/widgets/dialogs/aves_dialog.dart';
-import 'package:aves/widgets/dialogs/aves_selection_dialog.dart';
-import 'package:aves/widgets/filter_grids/album_pick.dart';
 import 'package:aves/widgets/map/map_page.dart';
 import 'package:aves/widgets/search/search_delegate.dart';
 import 'package:aves/widgets/stats/stats_page.dart';
@@ -39,7 +32,9 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:tuple/tuple.dart';
 
-class EntrySetActionDelegate with EntryEditorMixin, FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
+import '../common/action_mixins/entry_storage.dart';
+
+class EntrySetActionDelegate with FeedbackMixin, PermissionAwareMixin, SizeAwareMixin, EntryEditorMixin, EntryStorageMixin {
   bool isVisible(
     EntrySetAction action, {
     required AppMode appMode,
@@ -67,13 +62,13 @@ class EntrySetActionDelegate with EntryEditorMixin, FeedbackMixin, PermissionAwa
       // browsing or selecting
       case EntrySetAction.map:
       case EntrySetAction.stats:
+      case EntrySetAction.rescan:
         return appMode == AppMode.main;
       // selecting
       case EntrySetAction.share:
       case EntrySetAction.delete:
       case EntrySetAction.copy:
       case EntrySetAction.move:
-      case EntrySetAction.rescan:
       case EntrySetAction.toggleFavourite:
       case EntrySetAction.rotateCCW:
       case EntrySetAction.rotateCW:
@@ -111,13 +106,13 @@ class EntrySetActionDelegate with EntryEditorMixin, FeedbackMixin, PermissionAwa
         return true;
       case EntrySetAction.map:
       case EntrySetAction.stats:
+      case EntrySetAction.rescan:
         return (!isSelecting && hasItems) || (isSelecting && hasSelection);
       // selecting
       case EntrySetAction.share:
       case EntrySetAction.delete:
       case EntrySetAction.copy:
       case EntrySetAction.move:
-      case EntrySetAction.rescan:
       case EntrySetAction.toggleFavourite:
       case EntrySetAction.rotateCCW:
       case EntrySetAction.rotateCW:
@@ -156,6 +151,9 @@ class EntrySetActionDelegate with EntryEditorMixin, FeedbackMixin, PermissionAwa
       case EntrySetAction.stats:
         _goToStats(context);
         break;
+      case EntrySetAction.rescan:
+        _rescan(context);
+        break;
       // selecting
       case EntrySetAction.share:
         _share(context);
@@ -168,9 +166,6 @@ class EntrySetActionDelegate with EntryEditorMixin, FeedbackMixin, PermissionAwa
         break;
       case EntrySetAction.move:
         _move(context, moveType: MoveType.move);
-        break;
-      case EntrySetAction.rescan:
-        _rescan(context);
         break;
       case EntrySetAction.toggleFavourite:
         _toggleFavourite(context);
@@ -215,12 +210,12 @@ class EntrySetActionDelegate with EntryEditorMixin, FeedbackMixin, PermissionAwa
   }
 
   void _rescan(BuildContext context) {
-    final source = context.read<CollectionSource>();
     final selection = context.read<Selection<AvesEntry>>();
-    final selectedItems = _getExpandedSelectedItems(selection);
+    final collection = context.read<CollectionLens>();
+    final entries = (selection.isSelecting ? _getExpandedSelectedItems(selection) : collection.sortedEntries.toSet());
 
     final controller = AnalysisController(canStartService: true, force: true);
-    source.analyze(controller, entries: selectedItems);
+    collection.source.analyze(controller, entries: entries);
 
     selection.browse();
   }
@@ -268,7 +263,7 @@ class EntrySetActionDelegate with EntryEditorMixin, FeedbackMixin, PermissionAwa
 
     source.pauseMonitoring();
     final opId = mediaFileService.newOpId;
-    showOpReport<ImageOpEvent>(
+    await showOpReport<ImageOpEvent>(
       context: context,
       opStream: mediaFileService.delete(opId: opId, entries: selectedItems),
       itemCount: todoCount,
@@ -294,132 +289,10 @@ class EntrySetActionDelegate with EntryEditorMixin, FeedbackMixin, PermissionAwa
   }
 
   Future<void> _move(BuildContext context, {required MoveType moveType}) async {
-    final l10n = context.l10n;
     final selection = context.read<Selection<AvesEntry>>();
     final selectedItems = _getExpandedSelectedItems(selection);
-    final selectionDirs = selectedItems.map((e) => e.directory).whereNotNull().toSet();
-
-    final destinationAlbum = await pickAlbum(context: context, moveType: moveType);
-    if (destinationAlbum == null) return;
-    if (!await checkStoragePermissionForAlbums(context, {destinationAlbum})) return;
-
-    if (moveType == MoveType.move && !await checkStoragePermissionForAlbums(context, selectionDirs, entries: selectedItems)) return;
-
-    if (!await checkFreeSpaceForMove(context, selectedItems, destinationAlbum, moveType)) return;
-
-    // do not directly use selection when moving and post-processing items
-    // as source monitoring may remove obsolete items from the original selection
-    final todoItems = selectedItems.toSet();
-
-    final copy = moveType == MoveType.copy;
-    final todoCount = todoItems.length;
-    assert(todoCount > 0);
-
-    final destinationDirectory = Directory(destinationAlbum);
-    final names = [
-      ...todoItems.map((v) => '${v.filenameWithoutExtension}${v.extension}'),
-      // do not guard up front based on directory existence,
-      // as conflicts could be within moved entries scattered across multiple albums
-      if (await destinationDirectory.exists()) ...destinationDirectory.listSync().map((v) => pContext.basename(v.path)),
-    ];
-    final uniqueNames = names.toSet();
-    var nameConflictStrategy = NameConflictStrategy.rename;
-    if (uniqueNames.length < names.length) {
-      final value = await showDialog<NameConflictStrategy>(
-        context: context,
-        builder: (context) {
-          return AvesSelectionDialog<NameConflictStrategy>(
-            initialValue: nameConflictStrategy,
-            options: Map.fromEntries(NameConflictStrategy.values.map((v) => MapEntry(v, v.getName(context)))),
-            message: selectionDirs.length == 1 ? l10n.nameConflictDialogSingleSourceMessage : l10n.nameConflictDialogMultipleSourceMessage,
-            confirmationButtonLabel: l10n.continueButtonLabel,
-          );
-        },
-      );
-      if (value == null) return;
-      nameConflictStrategy = value;
-    }
-
-    final source = context.read<CollectionSource>();
-    source.pauseMonitoring();
-    final opId = mediaFileService.newOpId;
-    showOpReport<MoveOpEvent>(
-      context: context,
-      opStream: mediaFileService.move(
-        opId: opId,
-        entries: todoItems,
-        copy: copy,
-        destinationAlbum: destinationAlbum,
-        nameConflictStrategy: nameConflictStrategy,
-      ),
-      itemCount: todoCount,
-      onCancel: () => mediaFileService.cancelFileOp(opId),
-      onDone: (processed) async {
-        final successOps = processed.where((e) => e.success).toSet();
-        final movedOps = successOps.where((e) => !e.skipped).toSet();
-        await source.updateAfterMove(
-          todoEntries: todoItems,
-          copy: copy,
-          destinationAlbum: destinationAlbum,
-          movedOps: movedOps,
-        );
-        selection.browse();
-        source.resumeMonitoring();
-
-        // cleanup
-        if (moveType == MoveType.move) {
-          await storageService.deleteEmptyDirectories(selectionDirs);
-        }
-
-        final successCount = successOps.length;
-        if (successCount < todoCount) {
-          final count = todoCount - successCount;
-          showFeedback(context, copy ? l10n.collectionCopyFailureFeedback(count) : l10n.collectionMoveFailureFeedback(count));
-        } else {
-          final count = movedOps.length;
-          showFeedback(
-            context,
-            copy ? l10n.collectionCopySuccessFeedback(count) : l10n.collectionMoveSuccessFeedback(count),
-            count > 0
-                ? SnackBarAction(
-                    label: l10n.showButtonLabel,
-                    onPressed: () async {
-                      final highlightInfo = context.read<HighlightInfo>();
-                      final collection = context.read<CollectionLens>();
-                      var targetCollection = collection;
-                      if (collection.filters.any((f) => f is AlbumFilter)) {
-                        final filter = AlbumFilter(destinationAlbum, source.getAlbumDisplayName(context, destinationAlbum));
-                        // we could simply add the filter to the current collection
-                        // but navigating makes the change less jarring
-                        targetCollection = CollectionLens(
-                          source: collection.source,
-                          filters: collection.filters,
-                        )..addFilter(filter);
-                        unawaited(Navigator.pushReplacement(
-                          context,
-                          MaterialPageRoute(
-                            settings: const RouteSettings(name: CollectionPage.routeName),
-                            builder: (context) => CollectionPage(
-                              collection: targetCollection,
-                            ),
-                          ),
-                        ));
-                        final delayDuration = context.read<DurationsData>().staggeredAnimationPageTarget;
-                        await Future.delayed(delayDuration);
-                      }
-                      await Future.delayed(Durations.highlightScrollInitDelay);
-                      final newUris = movedOps.map((v) => v.newFields['uri'] as String?).toSet();
-                      final targetEntry = targetCollection.sortedEntries.firstWhereOrNull((entry) => newUris.contains(entry.uri));
-                      if (targetEntry != null) {
-                        highlightInfo.trackItem(targetEntry, highlightItem: targetEntry);
-                      }
-                    },
-                  )
-                : null,
-          );
-        }
-      },
-    );
+    await move(context, moveType: moveType, selectedItems: selectedItems);
+    selection.browse();
   }
 
   Future<void> _edit(
@@ -439,7 +312,7 @@ class EntrySetActionDelegate with EntryEditorMixin, FeedbackMixin, PermissionAwa
     final source = context.read<CollectionSource>();
     source.pauseMonitoring();
     var cancelled = false;
-    showOpReport<ImageOpEvent>(
+    await showOpReport<ImageOpEvent>(
       context: context,
       opStream: Stream.fromIterable(todoItems).asyncMap((entry) async {
         if (cancelled) {